backend/pyproject.toml
[project]
name = "todo-backend"
version = "0.1.0"
description = "Backend API for Todo application with authentication"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [{name = "Developer", email = "dev@example.com"}]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "sqlmodel>=0.0.22",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.1",
    "pydantic[email]>=2.10.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib>=1.7.0",
    "bcrypt>=4.0.0",
    "better-exceptions>=0.3.3",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.0.1",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "black>=24.10.0",
    "flake8>=7.1.1",
    "mypy>=1.13.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
dev = "src.main:dev"